+++ src/lv_font/lv_font_unscii_16.c
@@ -7,7 +7,7 @@
  ******************************************************************************/
 
 #ifndef LV_FONT_UNSCII_16
-#define LV_FONT_UNSCII_16 1
+    #define LV_FONT_UNSCII_16 1
 #endif
 
 #if LV_FONT_UNSCII_16
@@ -583,8 +583,7 @@ static const lv_font_fmt_txt_glyph_dsc_t glyph_dsc[] = {
 
 
 /*Collect the unicode lists and glyph_id offsets*/
-static const lv_font_fmt_txt_cmap_t cmaps[] =
-{
+static const lv_font_fmt_txt_cmap_t cmaps[] = {
     {
         .range_start = 32, .range_length = 96, .glyph_id_start = 1,
         .unicode_list = NULL, .glyph_id_ofs_list = NULL, .list_length = 0, .type = LV_FONT_FMT_TXT_CMAP_FORMAT0_TINY