+++ src/lv_widgets/lv_roller.c
@@ -181,7 +181,7 @@ void lv_roller_set_options(lv_obj_t * roller, const char * options, lv_roller_mo
         lv_label_set_text(label, options);
     }
     else {
-        ext->mode = LV_ROLLER_MODE_INIFINITE;
+        ext->mode = LV_ROLLER_MODE_INFINITE;
 
         size_t opt_len = strlen(options) + 1; /*+1 to add '\n' after option lists*/
         char * opt_extra = _lv_mem_buf_get(opt_len * LV_ROLLER_INF_PAGES);
@@ -244,7 +244,7 @@ void lv_roller_set_selected(lv_obj_t * roller, uint16_t sel_opt, lv_anim_enable_
     lv_roller_ext_t * ext = lv_obj_get_ext_attr(roller);
 
     /*In infinite mode interpret the new ID relative to the currently visible "page"*/
-    if(ext->mode == LV_ROLLER_MODE_INIFINITE) {
+    if(ext->mode == LV_ROLLER_MODE_INFINITE) {
         int32_t sel_opt_signed = sel_opt;
         uint16_t page = ext->sel_opt_id / LV_ROLLER_INF_PAGES;
 
@@ -304,7 +304,7 @@ uint16_t lv_roller_get_selected(const lv_obj_t * roller)
     LV_ASSERT_OBJ(roller, LV_OBJX_NAME);
 
     lv_roller_ext_t * ext = lv_obj_get_ext_attr(roller);
-    if(ext->mode == LV_ROLLER_MODE_INIFINITE) {
+    if(ext->mode == LV_ROLLER_MODE_INFINITE) {
         uint16_t real_id_cnt = ext->option_cnt / LV_ROLLER_INF_PAGES;
         return ext->sel_opt_id % real_id_cnt;
     }
@@ -357,7 +357,7 @@ uint16_t lv_roller_get_option_cnt(const lv_obj_t * roller)
     LV_ASSERT_OBJ(roller, LV_OBJX_NAME);
 
     lv_roller_ext_t * ext = lv_obj_get_ext_attr(roller);
-    if(ext->mode == LV_ROLLER_MODE_INIFINITE) {
+    if(ext->mode == LV_ROLLER_MODE_INFINITE) {
         return ext->option_cnt / LV_ROLLER_INF_PAGES;
     }
     else {
@@ -896,7 +896,7 @@ static void inf_normalize(lv_obj_t * roller)
 {
     lv_roller_ext_t * ext  = lv_obj_get_ext_attr(roller);
 
-    if(ext->mode == LV_ROLLER_MODE_INIFINITE) {
+    if(ext->mode == LV_ROLLER_MODE_INFINITE) {
         uint16_t real_id_cnt = ext->option_cnt / LV_ROLLER_INF_PAGES;
         ext->sel_opt_id = ext->sel_opt_id % real_id_cnt;
         ext->sel_opt_id += (LV_ROLLER_INF_PAGES / 2) * real_id_cnt; /*Select the middle page*/